修改主键泛型为Int类型

+++ src/main/java/com/lord/advertis/dao/BrowseDao.java
@@ -3,6 +3,6 @@ package com.lord.advertis.dao;
 import com.lord.advertis.base.BaseMapper;
 import com.lord.advertis.domain.Browse;
 
-public interface BrowseDao extends BaseMapper<Browse, Long>{
+public interface BrowseDao extends BaseMapper<Browse, Integer>{
 
 }
